fix onekey fee UI
