+++ packages/kit/src/views/Swap/components/SwapProviderInfoItem.tsx
@@ -1,9 +1,14 @@
+import { useCallback, useMemo } from 'react';
+
 import { useIntl } from 'react-intl';
 
+import type { IImageSourceProps } from '@onekeyhq/components';
 import {
   Badge,
   Icon,
+  IconButton,
   Image,
+  Popover,
   SizableText,
   Skeleton,
   Stack,
@@ -22,6 +27,14 @@ interface ISwapProviderInfoItemProps {
   onPress?: () => void;
   isLoading?: boolean;
 }
+interface IProtocolFeeInfo {
+  name: string;
+  fee: number;
+  color: string;
+  icon: IImageSourceProps['source'];
+  maxFee: number;
+}
+
 const SwapProviderInfoItem = ({
   fromToken,
   isBest,
@@ -33,11 +46,113 @@ const SwapProviderInfoItem = ({
   isLoading,
 }: ISwapProviderInfoItemProps) => {
   const intl = useIntl();
+  const protocolFeeInfoList: IProtocolFeeInfo[] = useMemo(
+    () => [
+      {
+        maxFee: 0.875,
+        name: 'metamask',
+        color: '#F5841F',
+        icon: {
+          uri: 'https://uni.onekey-asset.com/static/logo/metamasklogo.png',
+        },
+        fee: 0.875,
+      },
+      {
+        maxFee: 0.875,
+        name: 'phantom',
+        fee: 0.85,
+        color: '#AB9FF2',
+
+        icon: {
+          uri: 'https://uni.onekey-asset.com/static/logo/Phantom.png',
+        },
+      },
+      {
+        maxFee: 0.875,
+        name: 'zerion',
+        fee: 0.8,
+        color: '#2461ED',
+
+        icon: {
+          uri: 'https://uni.onekey-asset.com/static/logo/zerionlogo.png',
+        },
+      },
+      {
+        maxFee: 0.875,
+        name: 'oneKey',
+        fee: 0.3,
+        // color: '#202020',
+        color: '$bgInverse',
+        icon: require('@onekeyhq/kit/assets/logo.png'),
+      },
+    ],
+    [],
+  );
+  const renderProtocolFeeListItem = useCallback(
+    (item: IProtocolFeeInfo) => (
+      <XStack gap="$3" alignItems="center">
+        <Stack w={20} h={20}>
+          <Image source={item.icon} w={16} h={16} />
+        </Stack>
+        <Stack flex={1}>
+          <Stack
+            backgroundColor={item.color}
+            borderRadius="$full"
+            width={`${item.maxFee > 0 ? (item.fee / item.maxFee) * 100 : 0}%`}
+            height="$1"
+          />
+        </Stack>
+        <SizableText
+          size="$bodySm"
+          color={item.name === 'oneKey' ? '$textSuccess' : '$text'}
+          textAlign="right"
+        >
+          {item.fee}%
+        </SizableText>
+      </XStack>
+    ),
+    [],
+  );
   return (
     <XStack justifyContent="space-between" alignItems="center">
-      <SizableText size="$bodyMd" color="$textSubdued" userSelect="none">
-        {intl.formatMessage({ id: ETranslations.swap_page_provider_provider })}
-      </SizableText>
+      <XStack>
+        <SizableText
+          size="$bodyMd"
+          color="$textSubdued"
+          userSelect="none"
+          mr="$1"
+        >
+          {intl.formatMessage({
+            id: ETranslations.swap_page_provider_provider,
+          })}
+        </SizableText>
+        <Popover
+          title={intl.formatMessage({
+            id: ETranslations.provider_ios_popover_onekey_fee,
+          })}
+          renderTrigger={
+            <IconButton
+              variant="tertiary"
+              size="small"
+              icon="InfoCircleOutline"
+            />
+          }
+          renderContent={
+            <Stack gap="$5" p="$4">
+              <SizableText size="$bodyMd" color="$textSubdued">
+                {intl.formatMessage({
+                  id: ETranslations.provider_ios_popover_onekey_fee_content,
+                })}
+              </SizableText>
+              <Stack gap="$2">
+                {protocolFeeInfoList.map((item) =>
+                  renderProtocolFeeListItem(item),
+                )}
+              </Stack>
+            </Stack>
+          }
+        />
+      </XStack>
 
       {isLoading ? (
         <Stack py="$1">
