+++ packages/shared/src/locale/json/bn.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "যোগাযোগ তথ্য মুছুন",
   "address_book_edit_address.title": "ঠিকানা সম্পাদনা করুন",
   "address_book_select.title": "যোগাযোগের বিকল্প নির্বাচন করুন",
-  "address_input.contract_popover": "আপনি একটি চুক্তির ঠিকানায় পাঠাচ্ছেন, নিয়মিত ব্যবহারকারীর ঠিকানা নয়৷ ঠিকানাটি দুবার চেক করুন এবং ঝুঁকির জন্য দেখুন।",
+  "address_input.contract_popover": "আপনি একটি চুক্তির ঠিকানায় পাঠাচ্ছেন, একটি সাধারণ ব্যবহারকারীর ঠিকানায় নয়। চুক্তির ঠিকানায় ভুল স্থানান্তর সম্পদের ক্ষতির কারণ হতে পারে। ঠিকানাটি আবার পরীক্ষা করুন এবং ঝুঁকির জন্য সতর্ক থাকুন।",
   "address_input.first_transfer_popover": "প্রথমে {network} এ এই ঠিকানায় স্থানান্তর করুন। ঠিকানাটি দুবার চেক করুন এবং ঝুঁকির জন্য দেখুন।",
-  "address_input.transferred_popover": "এই ঠিকানায় আগের স্থানান্তর সনাক্ত করা হয়েছে. দয়া করে এগিয়ে যান, তবে সতর্ক থাকুন।",
+  "address_input.transferred_popover": "এই ঠিকানায় পূর্ববর্তী স্থানান্তর সনাক্ত করা হয়েছে। এটি উদ্দেশ্যপ্রণোদিত ঠিকানা কিনা তা নিশ্চিত করতে দ্বিগুণ যাচাই করতে ভুলবেন না।",
   "address_type": "ঠিকানার ধরণ",
   "approve_edit.approve_amount": "পরিমাণ অনুমোদন করুন",
   "approve_edit.less_than_swap": "ভাতা অদলবদলের পরিমাণের চেয়ে কম হতে পারে না",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "OneKey v4 এ ডাউনগ্রেড করবেন না। ডাউনগ্রেড করলে অপরিবর্তনীয় ডাটা হারানোর ঝুঁকি রয়েছে।",
   "downgrade_warning_title": "গুরুত্বপূর্ণ: ডাউনগ্রেড করবেন না",
   "earn.24h_earnings": "24 ঘন্টার উপার্জন",
-  "earn.24h_earnings_tooltip": "আপনার সক্রিয় পরিমাণের উপর ভিত্তি করে গত ২৪ ঘন্টার আনুমানিক উপার্জন।",
+  "earn.24h_earnings_tooltip": "আপনার সক্রিয় পরিমাণের উপর ভিত্তি করে আনুমানিক ২৪ ঘণ্টার উপার্জন।",
   "earn.active": "সক্রিয়",
   "earn.annually": "প্রতিবছরে",
   "earn.approaching_staking_cap": "আমরা যেহেতু স্টেকিং সীমার কাছাকাছি চলে এসেছি, দয়া করে আপনার স্টেকিং ফলাফল চেক করুন। আপনাকে পুনরায় স্টেক করতে হবে যদি ওভারফ্লো হিসাবে চিহ্নিত করা হয়।",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "অদলবদল করার আগে টোকেন অনুমোদন করতে হবে",
   "provider.ios_popover.approval_require_title": "অনুমোদন প্রয়োজন",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "উদ্ধৃতি অন্তর্গত একটি 0.3% OneKey ফি রয়েছে যা নিশ্চিত করে যে আমরা উচ্চ মানের সেবা এবং সমন্বিত অভিজ্ঞতা প্রদান করি।",
+  "provider.ios_popover.onekey_fee_content": "কোনও লুকানো ফি নেই। OneKey একটি 0.3% ফি চার্জ করে, যা ইতিমধ্যেই আপনার কোটে অন্তর্ভুক্ত, এবং লেনদেন মোড়ানোর জন্য কোনও চার্জ নেই। এটি আমাদের একটি উচ্চ-মানের, নিরবচ্ছিন্ন অভিজ্ঞতা প্রদান করতে সহায়তা করে।",
   "provider.ios_popover.order_info_title": "অর্ডারের তথ্য",
   "provider.ios_popover.title": "তথ্য",
   "provider.label_anti_mev": "এমইভি বিরোধী",
