+++ packages/shared/src/locale/json/de.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "Kontakt löschen",
   "address_book_edit_address.title": "Adresse bearbeiten",
   "address_book_select.title": "Kontakt auswählen",
-  "address_input.contract_popover": "Sie senden an eine Vertragsadresse, nicht an eine normale Benutzeradresse. Überprüfen Sie die Adresse noch einmal und achten Sie auf Risiken.",
+  "address_input.contract_popover": "Sie senden an eine Vertragsadresse, nicht an eine normale Benutzeradresse. Falsche Überweisungen an eine Vertragsadresse können zu einem Verlust von Vermögenswerten führen. Überprüfen Sie die Adresse sorgfältig und achten Sie auf Risiken.",
   "address_input.first_transfer_popover": "Überweisen Sie zuerst an diese Adresse im {network} . Überprüfen Sie die Adresse noch einmal und achten Sie auf Risiken.",
-  "address_input.transferred_popover": "Es wurde eine vorherige Überweisung an diese Adresse erkannt. Bitte fahren Sie fort, aber bleiben Sie vorsichtig.",
+  "address_input.transferred_popover": "Frühere Überweisungen an diese Adresse wurden erkannt. Vergessen Sie nicht, doppelt zu überprüfen, um sicherzustellen, dass es sich um die beabsichtigte Adresse handelt.",
   "address_type": "Adresstyp",
   "approve_edit.approve_amount": "Betrag genehmigen",
   "approve_edit.less_than_swap": "Der Freibetrag darf nicht geringer sein als der Swap-Betrag",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "Downgraden Sie NICHT auf OneKey v4. Ein Downgrade führt zu irreversiblen Datenverlusten.",
   "downgrade_warning_title": "Wichtig: Nicht herabstufen",
   "earn.24h_earnings": "24h Einnahmen",
-  "earn.24h_earnings_tooltip": "Geschätzte Einnahmen der letzten 24 Stunden basierend auf Ihrem aktiven Betrag.",
+  "earn.24h_earnings_tooltip": "Geschätzte 24-Stunden-Einnahmen basierend auf Ihrem aktivierten Betrag.",
   "earn.active": "Aktiv",
   "earn.annually": "Jährlich",
   "earn.approaching_staking_cap": "Da wir uns der Staking-Obergrenze nähern, überprüfen Sie bitte Ihre Staking-Ergebnisse. Sie müssen erneut staken, wenn Sie als Überlauf markiert sind.",
@@ -807,7 +807,7 @@
   "global.current": "Aktuell",
   "global.customize_nonce": "Nonce anpassen",
   "global.customize_nonce_desc": "Aktivieren Sie diese Option, um die Nonce (Transaktionsnummer) zu ändern und Ihre Transaktionsreihenfolge bei Bestätigung > Erweiterte Einstellungen zu steuern. Dies ist eine erweiterte Funktion, verwenden Sie sie mit Vorsicht.",
-  "global.dapp_interaction": "Dapp-Interaktion",
+  "global.dapp_interaction": "dApp-Interaktion",
   "global.dark": "Dunkel",
   "global.date": "Datum",
   "global.date_today": "Heute",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "Token muss vor dem Tausch genehmigt werden",
   "provider.ios_popover.approval_require_title": "Zustimmung erforderlich",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "Das Angebot beinhaltet eine 0,3% OneKey-Gebühr, um sicherzustellen, dass wir einen hochwertigen Service und ein nahtloses Erlebnis bieten",
+  "provider.ios_popover.onekey_fee_content": "Keine versteckten Gebühren. OneKey erhebt eine Gebühr von 0,3 %, die bereits in Ihrem Angebot enthalten ist, ohne Gebühren für das Wrapping von Transaktionen. Dies unterstützt uns dabei, ein hochwertiges, nahtloses Erlebnis zu bieten.",
   "provider.ios_popover.order_info_title": "Bestellinformationen",
   "provider.ios_popover.title": "Information",
   "provider.label_anti_mev": "Anti-MEV",
