+++ packages/shared/src/locale/json/en_US.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "Delete Contact",
   "address_book_edit_address.title": "Edit Address",
   "address_book_select.title": "Select Contact",
-  "address_input.contract_popover": "You are sending to a contract address, not a regular user address. Double-check the address and watch for risks.",
+  "address_input.contract_popover": "You are sending to a contract address, not a regular user address. Incorrect transfers to a contract address may result in asset loss. Double-check the address and watch for risks.",
   "address_input.first_transfer_popover": "First transfer to this address on {network}. Double-check the address and watch for risks.",
-  "address_input.transferred_popover": "Detected previous transfer to this address",
+  "address_input.transferred_popover": "Previous transfers to this address have been detected. Don’t forget to double-check to ensure it is the intended address.",
   "address_type": "Address type",
   "approve_edit.approve_amount": "Approve amount",
   "approve_edit.less_than_swap": "Allowance can't be less than the swap amount",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "Do NOT downgrade to OneKey v4. Downgrading will lead to irreversible data loss.",
   "downgrade_warning_title": "Important: Do not downgrade",
   "earn.24h_earnings": "24h earnings",
-  "earn.24h_earnings_tooltip": "Estimated earnings over the past 24 hours based on your active amount.",
+  "earn.24h_earnings_tooltip": "Estimated 24 hours earnings based on your activated amount.",
   "earn.active": "Active",
   "earn.annually": "Annually",
   "earn.approaching_staking_cap": "As we’re approaching the staking cap, please check your staking results. You’ll need to restake if marked as overflow.",
@@ -807,7 +807,7 @@
   "global.current": "Current",
   "global.customize_nonce": "Customize nonce",
   "global.customize_nonce_desc": "Turn this on to change the nonce (transaction number) and control your transaction order on confirmation > advance settings. This is an advanced feature, use cautiously.",
-  "global.dapp_interaction": "Dapp Interaction",
+  "global.dapp_interaction": "dApp Interaction",
   "global.dark": "Dark",
   "global.date": "Date",
   "global.date_today": "Today",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "Need to approve token before swap",
   "provider.ios_popover.approval_require_title": "Approval Required",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "Quote includes a 0.3% OneKey fee to ensure we provide high-quality service and the seamless experience",
+  "provider.ios_popover.onekey_fee_content": "No hidden fees. OneKey charge a 0.3% fee, already included in your quote, with no charges for wrapping transactions. This supports us in providing a high-quality, seamless experience.",
   "provider.ios_popover.order_info_title": "Order Info",
   "provider.ios_popover.title": "Information",
   "provider.label_anti_mev": "Anti-MEV",
