+++ packages/shared/src/locale/json/es.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "Eliminar Contacto",
   "address_book_edit_address.title": "Editar dirección",
   "address_book_select.title": "Seleccionar Contacto",
-  "address_input.contract_popover": "Estás enviando a una dirección de contrato, no a una dirección de usuario normal. Verifica la dirección y ten cuidado con los riesgos.",
+  "address_input.contract_popover": "Estás enviando a una dirección de contrato, no a una dirección de usuario regular. Las transferencias incorrectas a una dirección de contrato pueden resultar en la pérdida de activos. Verifica la dirección dos veces y ten cuidado con los riesgos.",
   "address_input.first_transfer_popover": "Primera transferencia a esta dirección en {network} . Verifique la dirección y esté atento a los riesgos.",
-  "address_input.transferred_popover": "Se detectó una transferencia anterior a esta dirección. Continúe, pero tenga cuidado.",
+  "address_input.transferred_popover": "Se han detectado transferencias anteriores a esta dirección. No olvides verificar dos veces para asegurarte de que es la dirección deseada.",
   "address_type": "Tipo de dirección",
   "approve_edit.approve_amount": "Aprobar monto",
   "approve_edit.less_than_swap": "La asignación no puede ser menor que el monto del swap",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "No DEGRADES a OneKey v4. La degradación provocará una pérdida de datos irreversible.",
   "downgrade_warning_title": "Importante: No haga un downgrade",
   "earn.24h_earnings": "Ganancias de 24h",
-  "earn.24h_earnings_tooltip": "Ganancias estimadas de las últimas 24 horas basadas en tu monto activo.",
+  "earn.24h_earnings_tooltip": "Ganancias estimadas de 24 horas basadas en tu monto activado.",
   "earn.active": "Activo",
   "earn.annually": "Anualmente",
   "earn.approaching_staking_cap": "A medida que nos acercamos al límite de staking, por favor revisa tus resultados de staking. Necesitarás volver a hacer staking si se marca como exceso.",
@@ -807,7 +807,7 @@
   "global.current": "Actual",
   "global.customize_nonce": "Personalizar nonce",
   "global.customize_nonce_desc": "Active esta opción para cambiar el nonce (número de transacción) y controlar el orden de sus transacciones en la confirmación > configuración avanzada. Esta es una función avanzada, utilícela con precaución.",
-  "global.dapp_interaction": "Interacción con Dapp",
+  "global.dapp_interaction": "Interacción con dApp",
   "global.dark": "Oscuro",
   "global.date": "Fecha",
   "global.date_today": "Hoy",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "Es necesario aprobar el token antes del intercambio",
   "provider.ios_popover.approval_require_title": "Aprobación Requerida",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "La cotización incluye una tarifa de OneKey del 0.3% para garantizar que proporcionamos un servicio de alta calidad y una experiencia sin problemas",
+  "provider.ios_popover.onekey_fee_content": "Sin tarifas ocultas. OneKey cobra una tarifa del 0.3%, ya incluida en su cotización, sin cargos por transacciones de envoltura. Esto nos ayuda a proporcionar una experiencia de alta calidad y sin interrupciones.",
   "provider.ios_popover.order_info_title": "Información del Pedido",
   "provider.ios_popover.title": "Información",
   "provider.label_anti_mev": "Anti-MEV",
