+++ packages/shared/src/locale/json/fr_FR.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "Supprimer le contact",
   "address_book_edit_address.title": "Modifier l'adresse",
   "address_book_select.title": "Sélectionner un contact",
-  "address_input.contract_popover": "Vous envoyez à une adresse de contrat et non à une adresse d'utilisateur ordinaire. Vérifiez l'adresse et surveillez les risques.",
+  "address_input.contract_popover": "Vous envoyez à une adresse de contrat, et non à une adresse utilisateur régulière. Les transferts incorrects vers une adresse de contrat peuvent entraîner une perte d'actifs. Vérifiez l'adresse à deux reprises et soyez attentif aux risques.",
   "address_input.first_transfer_popover": "Effectuez d'abord le transfert vers cette adresse sur {network} . Vérifiez l'adresse et faites attention aux risques.",
-  "address_input.transferred_popover": "Un transfert précédent a été détecté vers cette adresse. Veuillez continuer, mais restez prudent.",
+  "address_input.transferred_popover": "Des transferts précédents vers cette adresse ont été détectés. N'oubliez pas de vérifier deux fois pour vous assurer qu'il s'agit de l'adresse souhaitée.",
   "address_type": "Type d'adresse",
   "approve_edit.approve_amount": "Approuver le montant",
   "approve_edit.less_than_swap": "L'allocation ne peut pas être inférieure au montant du swap",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "Ne PAS rétrograder à OneKey v4. La rétrogradation entraînera une perte de données irréversible.",
   "downgrade_warning_title": "Important : Ne pas rétrograder",
   "earn.24h_earnings": "Gains sur 24h",
-  "earn.24h_earnings_tooltip": "Gains estimés au cours des dernières 24 heures en fonction de votre montant actif.",
+  "earn.24h_earnings_tooltip": "Gains estimés sur 24 heures en fonction de votre montant activé.",
   "earn.active": "Actif",
   "earn.annually": "Annuellement",
   "earn.approaching_staking_cap": "Alors que nous approchons du plafond de staking, veuillez vérifier vos résultats de staking. Vous devrez restaker si vous êtes marqué comme débordement.",
@@ -807,7 +807,7 @@
   "global.current": "Actuel",
   "global.customize_nonce": "Personnaliser le nonce",
   "global.customize_nonce_desc": "Activez cette option pour modifier le nonce (numéro de transaction) et contrôler l'ordre de vos transactions lors de la confirmation > paramètres avancés. Il s'agit d'une fonctionnalité avancée, à utiliser avec prudence.",
-  "global.dapp_interaction": "Interaction Dapp",
+  "global.dapp_interaction": "Interaction dApp",
   "global.dark": "Dark",
   "global.date": "Date",
   "global.date_today": "Aujourd'hui",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "Il faut approuver le jeton avant l'échange",
   "provider.ios_popover.approval_require_title": "Approbation requise",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "Le devis inclut des frais OneKey de 0,3% pour garantir que nous fournissons un service de haute qualité et une expérience sans faille",
+  "provider.ios_popover.onekey_fee_content": "Pas de frais cachés. OneKey prélève des frais de 0,3 %, déjà inclus dans votre devis, sans frais pour les transactions de wrapping. Cela nous aide à offrir une expérience de haute qualité et sans accroc.",
   "provider.ios_popover.order_info_title": "Informations de Commande",
   "provider.ios_popover.title": "Information",
   "provider.label_anti_mev": "Anti-MEV",
