+++ packages/shared/src/locale/json/hi_IN.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "संपर्क हटाएं",
   "address_book_edit_address.title": "पता संशोधित करें",
   "address_book_select.title": "संपर्क चुनें",
-  "address_input.contract_popover": "आप किसी नियमित उपयोगकर्ता के पते पर नहीं, बल्कि अनुबंध पते पर भेज रहे हैं। पते की दोबारा जाँच करें और जोखिमों पर नज़र रखें।",
+  "address_input.contract_popover": "आप एक अनुबंध पते पर भेज रहे हैं, न कि एक सामान्य उपयोगकर्ता पते पर। अनुबंध पते पर गलत स्थानांतरण से संपत्ति की हानि हो सकती है। पते की दोबारा जांच करें और जोखिमों के लिए सतर्क रहें।",
   "address_input.first_transfer_popover": "सबसे पहले {network} पर इस पते पर स्थानांतरण करें। पते की दोबारा जांच करें और जोखिमों पर नज़र रखें।",
-  "address_input.transferred_popover": "इस पते पर पहले भी स्थानांतरण का पता चला है। कृपया आगे बढ़ें, लेकिन सावधान रहें।",
+  "address_input.transferred_popover": "इस पते पर पहले के स्थानांतरण का पता चला है। यह सुनिश्चित करने के लिए दोबारा जांच करना न भूलें कि यह इच्छित पता है।",
   "address_type": "पता प्रकार",
   "approve_edit.approve_amount": "राशि स्वीकृत करें",
   "approve_edit.less_than_swap": "भत्ता स्वैप राशि से कम नहीं हो सकता",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "OneKey v4 के लिए अवनति न करें। अवनति करने से अपरिवर्तनीय डेटा हानि होगी।",
   "downgrade_warning_title": "महत्वपूर्ण: अवनति न करें",
   "earn.24h_earnings": "24 घंटे की कमाई",
-  "earn.24h_earnings_tooltip": "आपकी सक्रिय राशि के आधार पर पिछले 24 घंटों की अनुमानित कमाई।",
+  "earn.24h_earnings_tooltip": "आपके सक्रिय राशि के आधार पर अनुमानित 24 घंटे की कमाई।",
   "earn.active": "सक्रिय",
   "earn.annually": "प्रतिवर्ष",
   "earn.approaching_staking_cap": "जैसा कि हम स्टेकिंग की सीमा के नजदीक आ रहे हैं, कृपया अपने स्टेकिंग परिणामों की जांच करें। यदि आपको ओवरफ्लो के रूप में चिह्नित किया गया है, तो आपको पुनः स्टेक करने की आवश्यकता होगी।",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "स्वैप से पहले टोकन को अनुमोदित करना आवश्यक है",
   "provider.ios_popover.approval_require_title": "अनुमोदन आवश्यक",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "उद्धरण में 0.3% OneKey शुल्क शामिल है ताकि हम उच्च-गुणवत्ता वाली सेवा और सहज अनुभव प्रदान कर सकें।",
+  "provider.ios_popover.onekey_fee_content": "कोई छुपे हुए शुल्क नहीं। OneKey 0.3% शुल्क लेता है, जो पहले से ही आपके कोटेशन में शामिल है, और लेन-देन को लपेटने के लिए कोई शुल्क नहीं है। यह हमें उच्च-गुणवत्ता, सहज अनुभव प्रदान करने में सहायता करता है।",
   "provider.ios_popover.order_info_title": "आदेश की जानकारी",
   "provider.ios_popover.title": "सूचना",
   "provider.label_anti_mev": "एंटी-एमईवी",
