+++ packages/shared/src/locale/json/id.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "Hapus Kontak",
   "address_book_edit_address.title": "Sunting Alamat",
   "address_book_select.title": "Pilih Kontak",
-  "address_input.contract_popover": "Anda mengirim ke alamat kontrak, bukan alamat pengguna biasa. Periksa kembali alamat tersebut dan waspadai risikonya.",
+  "address_input.contract_popover": "Anda mengirim ke alamat kontrak, bukan alamat pengguna biasa. Transfer yang salah ke alamat kontrak dapat mengakibatkan kehilangan aset. Periksa kembali alamatnya dan waspadai risiko.",
   "address_input.first_transfer_popover": "Pertama-tama, transfer ke alamat ini di {network} . Periksa kembali alamatnya dan waspadai risikonya.",
-  "address_input.transferred_popover": "Telah terdeteksi adanya transfer sebelumnya ke alamat ini. Silakan lanjutkan, tetapi tetaplah berhati-hati.",
+  "address_input.transferred_popover": "Transfer sebelumnya ke alamat ini telah terdeteksi. Jangan lupa untuk memeriksa kembali untuk memastikan bahwa ini adalah alamat yang dimaksud.",
   "address_type": "Jenis alamat",
   "approve_edit.approve_amount": "Menyetujui jumlah",
   "approve_edit.less_than_swap": "Tunjangan tidak boleh kurang dari jumlah swap",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "Jangan menurunkan versi ke OneKey v4. Penurunan versi akan menyebabkan kehilangan data yang tidak dapat dikembalikan.",
   "downgrade_warning_title": "Penting: Jangan menurunkan versi",
   "earn.24h_earnings": "Pendapatan 24 jam",
-  "earn.24h_earnings_tooltip": "Perkiraan penghasilan selama 24 jam terakhir berdasarkan jumlah aktif Anda.",
+  "earn.24h_earnings_tooltip": "Perkiraan penghasilan 24 jam berdasarkan jumlah yang telah diaktifkan.",
   "earn.active": "Aktif",
   "earn.annually": "Setiap tahun",
   "earn.approaching_staking_cap": "Karena kita mendekati batas staking, silakan periksa hasil staking Anda. Anda perlu melakukan staking ulang jika ditandai sebagai overflow.",
@@ -807,7 +807,7 @@
   "global.current": "Saat ini",
   "global.customize_nonce": "Sesuaikan nonce",
   "global.customize_nonce_desc": "Aktifkan ini untuk mengubah nonce (nomor transaksi) dan mengontrol urutan transaksi Anda pada konfirmasi > pengaturan lanjutan. Ini adalah fitur lanjutan, gunakan dengan hati-hati.",
-  "global.dapp_interaction": "Interaksi Dapp",
+  "global.dapp_interaction": "Interaksi dApp",
   "global.dark": "Gelap",
   "global.date": "Tanggal",
   "global.date_today": "Hari ini",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "Perlu menyetujui token sebelum pertukaran",
   "provider.ios_popover.approval_require_title": "Persetujuan Diperlukan",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "Kutipan termasuk biaya OneKey sebesar 0,3% untuk memastikan kami memberikan layanan berkualitas tinggi dan pengalaman yang lancar",
+  "provider.ios_popover.onekey_fee_content": "Tidak ada biaya tersembunyi. OneKey mengenakan biaya 0,3%, yang sudah termasuk dalam penawaran Anda, tanpa biaya untuk transaksi wrapping. Ini mendukung kami dalam memberikan pengalaman yang berkualitas tinggi dan mulus.",
   "provider.ios_popover.order_info_title": "Info Pesanan",
   "provider.ios_popover.title": "Informasi",
   "provider.label_anti_mev": "Anti-MEV",
