+++ packages/shared/src/locale/json/it_IT.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "Elimina Contatto",
   "address_book_edit_address.title": "Modifica Indirizzo",
   "address_book_select.title": "Seleziona Contatto",
-  "address_input.contract_popover": "Stai inviando a un indirizzo contrattuale, non a un indirizzo utente normale. Controlla due volte l'indirizzo e fai attenzione ai rischi.",
+  "address_input.contract_popover": "Stai inviando a un indirizzo di contratto, non a un indirizzo utente normale. I trasferimenti errati a un indirizzo di contratto possono comportare la perdita di beni. Ricontrolla l'indirizzo e fai attenzione ai rischi.",
   "address_input.first_transfer_popover": "Per prima cosa trasferisci a questo indirizzo su {network} . Controlla attentamente l'indirizzo e fai attenzione ai rischi.",
-  "address_input.transferred_popover": "Rilevato un precedente trasferimento a questo indirizzo. Procedi, ma fai attenzione.",
+  "address_input.transferred_popover": "Trasferimenti precedenti a questo indirizzo sono stati rilevati. Non dimenticare di ricontrollare per assicurarti che sia l'indirizzo desiderato.",
   "address_type": "Tipo di indirizzo",
   "approve_edit.approve_amount": "Approva importo",
   "approve_edit.less_than_swap": "L'indennità non può essere inferiore all'importo dello swap",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "NON effettuare il downgrade a OneKey v4. Il downgrade porterà a una perdita di dati irreversibile.",
   "downgrade_warning_title": "Importante: Non effettuare il downgrade",
   "earn.24h_earnings": "Guadagni nelle 24 ore",
-  "earn.24h_earnings_tooltip": "Guadagni stimati nelle ultime 24 ore in base al tuo importo attivo.",
+  "earn.24h_earnings_tooltip": "Guadagni stimati sulle 24 ore in base al tuo importo attivato.",
   "earn.active": "Attivo",
   "earn.annually": "Annualmente",
   "earn.approaching_staking_cap": "Mentre ci avviciniamo al limite di staking, controlla i tuoi risultati di staking. Dovrai fare nuovamente staking se segnato come overflow.",
@@ -807,7 +807,7 @@
   "global.current": "Corrente",
   "global.customize_nonce": "Personalizza nonce",
   "global.customize_nonce_desc": "Attivalo per cambiare il nonce (numero di transazione) e controllare l'ordine della tua transazione alla conferma > impostazioni avanzate. Questa è una funzionalità avanzata, usala con cautela.",
-  "global.dapp_interaction": "Interazione Dapp",
+  "global.dapp_interaction": "Interazione dApp",
   "global.dark": "Buio",
   "global.date": "Data",
   "global.date_today": "Oggi",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "È necessario approvare il token prima dello scambio",
   "provider.ios_popover.approval_require_title": "Approvazione Richiesta",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "Il preventivo include una commissione OneKey dello 0,3% per garantire un servizio di alta qualità e un'esperienza senza problemi",
+  "provider.ios_popover.onekey_fee_content": "Nessun costo nascosto. OneKey applica una commissione dello 0,3%, già inclusa nel tuo preventivo, senza costi per le transazioni di wrapping. Questo ci aiuta a fornire un'esperienza di alta qualità e senza interruzioni.",
   "provider.ios_popover.order_info_title": "Informazioni sull'ordine",
   "provider.ios_popover.title": "Informazioni",
   "provider.label_anti_mev": "Anti-MEV",
