+++ packages/shared/src/locale/json/ja_JP.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "連絡先を削除",
   "address_book_edit_address.title": "住所を編集",
   "address_book_select.title": "連絡先を選択",
-  "address_input.contract_popover": "送信先は通常のユーザー アドレスではなく、契約アドレスです。アドレスを再確認し、リスクに注意してください。",
+  "address_input.contract_popover": "あなたは通常のユーザーアドレスではなく、コントラクトアドレスに送信しています。コントラクトアドレスへの誤った送金は資産の損失を招く可能性があります。アドレスを再確認し、リスクに注意してください。",
   "address_input.first_transfer_popover": "まず、 {network}上のこのアドレスに転送します。アドレスを再確認し、リスクに注意してください。",
-  "address_input.transferred_popover": "このアドレスへの以前の送金が検出されました。続行してください。ただし、注意してください。",
+  "address_input.transferred_popover": "このアドレスへの以前の送金が検出されました。意図したアドレスであることを確認するために、必ず再確認してください。",
   "address_type": "住所タイプ",
   "approve_edit.approve_amount": "金額を承認",
   "approve_edit.less_than_swap": "手当はスワップ金額より少なくすることはできません",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "OneKey v4にダウングレードしないでください。ダウングレードすると、データの損失が不可逆的になります。",
   "downgrade_warning_title": "重要：ダウングレードしないでください",
   "earn.24h_earnings": "24 時間の収益",
-  "earn.24h_earnings_tooltip": "あなたの有効な数量に基づく過去24時間の推定収益。",
+  "earn.24h_earnings_tooltip": "あなたのアクティベートされた金額に基づく24時間の推定収益。",
   "earn.active": "アクティブ",
   "earn.annually": "毎年",
   "earn.approaching_staking_cap": "ステーキングの上限に近づいているため、ステーキングの結果を確認してください。オーバーフローとマークされている場合は、再ステーキングが必要になります。",
@@ -807,7 +807,7 @@
   "global.current": "現在",
   "global.customize_nonce": "nonce をカスタマイズする",
   "global.customize_nonce_desc": "これをオンにすると、nonce (トランザクション番号) を変更し、確認時にトランザクションの順序を制御できます>詳細設定。これは高度な機能なので、慎重に使用してください。",
-  "global.dapp_interaction": "Dappの相互作用",
+  "global.dapp_interaction": "dApp インタラクション",
   "global.dark": "ダーク",
   "global.date": "日付",
   "global.date_today": "今日",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "スワップ前にトークンを承認する必要がある",
   "provider.ios_popover.approval_require_title": "承認が必要です",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "見積もりには、高品質なサービスとシームレスな体験を提供するための0.3%のOneKey手数料が含まれています",
+  "provider.ios_popover.onekey_fee_content": "隠れた手数料はありません。OneKeyは0.3%の手数料を請求しますが、これはすでにお見積もりに含まれており、トランザクションのラッピングには料金がかかりません。これにより、高品質でシームレスな体験を提供することができます。",
   "provider.ios_popover.order_info_title": "注文情報",
   "provider.ios_popover.title": "情報",
   "provider.label_anti_mev": "抗MEV",
