+++ packages/shared/src/locale/json/ko_KR.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "연락처 삭제",
   "address_book_edit_address.title": "주소 수정",
   "address_book_select.title": "연락처 선택",
-  "address_input.contract_popover": "일반 사용자 주소가 아닌 계약 주소로 보내고 있습니다. 주소를 다시 확인하고 위험에 주의하세요.",
+  "address_input.contract_popover": "계약 주소로 보내고 있으며, 일반 사용자 주소가 아닙니다. 계약 주소로의 잘못된 전송은 자산 손실을 초래할 수 있습니다. 주소를 두 번 확인하고 위험에 주의하십시오.",
   "address_input.first_transfer_popover": "{network} 에서 이 주소로 첫 번째 전송. 주소를 두 번 확인하고 위험에 주의하세요.",
-  "address_input.transferred_popover": "이 주소로의 이전 전송이 감지되었습니다. 계속 진행하세요. 하지만 조심하세요.",
+  "address_input.transferred_popover": "이 주소로의 이전 전송이 감지되었습니다. 의도한 주소인지 확인하기 위해 다시 한 번 확인하는 것을 잊지 마세요.",
   "address_type": "주소 유형",
   "approve_edit.approve_amount": "금액 승인",
   "approve_edit.less_than_swap": "허용량은 스왑 금액보다 적을 수 없습니다.",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "OneKey v4로 다운그레이드하지 마십시오. 다운그레이드하면 데이터 손실이 되돌릴 수 없게 됩니다.",
   "downgrade_warning_title": "중요: 다운그레이드하지 마세요",
   "earn.24h_earnings": "24 시간 수익",
-  "earn.24h_earnings_tooltip": "활성화된 금액을 기준으로 한 지난 24시간 동안의 예상 수익입니다.",
+  "earn.24h_earnings_tooltip": "활성화된 금액을 기준으로 한 24시간 예상 수익입니다.",
   "earn.active": "활성화됨",
   "earn.annually": "매년",
   "earn.approaching_staking_cap": "스테이킹 한도에 접근하고 있으므로, 스테이킹 결과를 확인해 주세요. 오버플로우로 표시된 경우 재스테이킹이 필요합니다.",
@@ -807,7 +807,7 @@
   "global.current": "현재",
   "global.customize_nonce": "nonce를 사용자 정의하세요",
   "global.customize_nonce_desc": "이것을 켜서 nonce(거래 번호)를 변경하고 확인 > 고급 설정에서 거래 순서를 제어합니다. 이것은 고급 기능이므로 신중하게 사용하세요.",
-  "global.dapp_interaction": "Dapp 상호 작용",
+  "global.dapp_interaction": "dApp 상호 작용",
   "global.dark": "다크",
   "global.date": "날짜",
   "global.date_today": "오늘",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "스왑 전 토큰 승인이 필요합니다.",
   "provider.ios_popover.approval_require_title": "승인 필요",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "\"견적에는 우수한 서비스와 원활한 경험을 제공하기 위해 0.3%의 OneKey 수수료가 포함되어 있습니다.\"",
+  "provider.ios_popover.onekey_fee_content": "숨겨진 수수료가 없습니다. OneKey는 0.3%의 수수료를 부과하며, 이는 이미 귀하의 견적에 포함되어 있으며 거래 포장에 대한 추가 요금은 없습니다. 이는 고품질의 원활한 경험을 제공하는 데 도움이 됩니다.",
   "provider.ios_popover.order_info_title": "주문 정보",
   "provider.ios_popover.title": "정보",
   "provider.label_anti_mev": "안티-MEV",
