+++ packages/shared/src/locale/json/pt.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "Excluir Contato",
   "address_book_edit_address.title": "Editar Endereço",
   "address_book_select.title": "Selecionar Contato",
-  "address_input.contract_popover": "Você está enviando para um endereço de contrato, não para um endereço de usuário regular. Verifique novamente o endereço e observe os riscos.",
+  "address_input.contract_popover": "Você está enviando para um endereço de contrato, não um endereço de usuário comum. Transferências incorretas para um endereço de contrato podem resultar em perda de ativos. Verifique o endereço duas vezes e fique atento aos riscos.",
   "address_input.first_transfer_popover": "Primeira transferência para este endereço em {network} . Verifique novamente o endereço e observe os riscos.",
-  "address_input.transferred_popover": "Transferência anterior detectada para este endereço. Por favor, prossiga, mas seja cauteloso.",
+  "address_input.transferred_popover": "Transferências anteriores para este endereço foram detectadas. Não se esqueça de verificar duas vezes para garantir que é o endereço pretendido.",
   "address_type": "Tipo de endereço",
   "approve_edit.approve_amount": "Aprovar valor",
   "approve_edit.less_than_swap": "O subsídio não pode ser menor que o valor do swap",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "NÃO faça o downgrade para o OneKey v4. O downgrade levará à perda irreversível de dados.",
   "downgrade_warning_title": "Importante: Não faça o downgrade",
   "earn.24h_earnings": "Ganhos de 24h",
-  "earn.24h_earnings_tooltip": "Ganhos estimados nas últimas 24 horas com base no seu montante ativo.",
+  "earn.24h_earnings_tooltip": "Ganhos estimados em 24 horas com base no seu montante ativado.",
   "earn.active": "Ativo",
   "earn.annually": "Anualmente",
   "earn.approaching_staking_cap": "À medida que nos aproximamos do limite de staking, por favor, verifique seus resultados de staking. Você precisará fazer o restake se estiver marcado como excesso.",
@@ -807,7 +807,7 @@
   "global.current": "Atual",
   "global.customize_nonce": "Personalizar nonce",
   "global.customize_nonce_desc": "Ative isso para alterar o nonce (número da transação) e controlar sua ordem de transação em confirmação > configurações avançadas. Este é um recurso avançado, use com cautela.",
-  "global.dapp_interaction": "Interação Dapp",
+  "global.dapp_interaction": "Interação dApp",
   "global.dark": "Escuro",
   "global.date": "Data",
   "global.date_today": "Hoje",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "É necessário aprovar o token antes da troca",
   "provider.ios_popover.approval_require_title": "Aprovação Necessária",
   "provider.ios_popover.onekey_fee": "Taxa OneKey",
-  "provider.ios_popover.onekey_fee_content": "A cotação inclui uma taxa de 0,3% da OneKey para garantir que oferecemos um serviço de alta qualidade e uma experiência sem interrupções",
+  "provider.ios_popover.onekey_fee_content": "Sem taxas ocultas. A OneKey cobra uma taxa de 0,3%, já incluída na sua cotação, sem cobranças para transações de wrapping. Isso nos apoia em fornecer uma experiência de alta qualidade e sem interrupções.",
   "provider.ios_popover.order_info_title": "Informações do Pedido",
   "provider.ios_popover.title": "Informação",
   "provider.label_anti_mev": "Anti-MEV",
