+++ packages/shared/src/locale/json/ru.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "Удалить контакт",
   "address_book_edit_address.title": "Изменить адрес",
   "address_book_select.title": "Выбрать контакт",
-  "address_input.contract_popover": "Вы отправляете на адрес контракта, а не на обычный адрес пользователя. Дважды проверьте адрес и следите за рисками.",
+  "address_input.contract_popover": "Вы отправляете на адрес контракта, а не на обычный адрес пользователя. Неправильные переводы на адрес контракта могут привести к потере активов. Дважды проверьте адрес и будьте внимательны к рискам.",
   "address_input.first_transfer_popover": "Сначала переведите на этот адрес в {network} . Дважды проверьте адрес и следите за рисками.",
-  "address_input.transferred_popover": "Обнаружен предыдущий перевод на этот адрес. Пожалуйста, продолжайте, но будьте осторожны.",
+  "address_input.transferred_popover": "Обнаружены предыдущие переводы на этот адрес. Не забудьте дважды проверить, чтобы убедиться, что это нужный адрес.",
   "address_type": "Тип адреса",
   "approve_edit.approve_amount": "Утвердить сумму",
   "approve_edit.less_than_swap": "Пособие не может быть меньше суммы свопа",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "НЕ понижайте версию до OneKey v4. Понижение версии приведет к необратимой потере данных.",
   "downgrade_warning_title": "Важно: не откатывайте версию",
   "earn.24h_earnings": "Заработок за 24 часа",
-  "earn.24h_earnings_tooltip": "Оценочные доходы за последние 24 часа на основе вашей активной суммы.",
+  "earn.24h_earnings_tooltip": "Оценка дохода за 24 часа на основе активированной суммы.",
   "earn.active": "Активный",
   "earn.annually": "Ежегодно",
   "earn.approaching_staking_cap": "Поскольку мы приближаемся к лимиту стейкинга, пожалуйста, проверьте результаты вашего стейкинга. Вам потребуется повторно стейкнуть, если отмечено как переполнение.",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "Необходимо одобрить токен перед обменом",
   "provider.ios_popover.approval_require_title": "Требуется утверждение",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "Цитата включает в себя комиссию OneKey в размере 0,3% для обеспечения высококачественного сервиса и бесшовного опыта",
+  "provider.ios_popover.onekey_fee_content": "Никаких скрытых комиссий. OneKey взимает комиссию в размере 0,3%, которая уже включена в вашу цену, без дополнительных сборов за обертывание транзакций. Это поддерживает нас в предоставлении высококачественного, бесшовного опыта.",
   "provider.ios_popover.order_info_title": "Информация о заказе",
   "provider.ios_popover.title": "Информация",
   "provider.label_anti_mev": "Анти-MEV",
