+++ packages/shared/src/locale/json/th_TH.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "ลบรายชื่อ",
   "address_book_edit_address.title": "แก้ไขที่อยู่",
   "address_book_select.title": "เลือกผู้ติดต่อ",
-  "address_input.contract_popover": "คุณกำลังส่งไปยังที่อยู่สัญญา ไม่ใช่ที่อยู่ผู้ใช้ทั่วไป ตรวจสอบที่อยู่อีกครั้งและระวังความเสี่ยง",
+  "address_input.contract_popover": "คุณกำลังส่งไปยังที่อยู่สัญญา ไม่ใช่ที่อยู่ผู้ใช้ทั่วไป การโอนที่ไม่ถูกต้องไปยังที่อยู่สัญญาอาจทำให้สูญเสียทรัพย์สินได้ ตรวจสอบที่อยู่อีกครั้งและระวังความเสี่ยง",
   "address_input.first_transfer_popover": "โอนไปยังที่อยู่นี้ใน {network} ก่อน ตรวจสอบที่อยู่อีกครั้งและระวังความเสี่ยง",
-  "address_input.transferred_popover": "ตรวจพบการโอนก่อนหน้านี้ไปยังที่อยู่นี้ โปรดดำเนินการต่อ แต่โปรดใช้ความระมัดระวัง",
+  "address_input.transferred_popover": "ตรวจพบการโอนก่อนหน้านี้ไปยังที่อยู่นี้ อย่าลืมตรวจสอบให้แน่ใจว่าเป็นที่อยู่ที่ต้องการ",
   "address_type": "ประเภทที่อยู่",
   "approve_edit.approve_amount": "จำนวนเงินอนุมัติ",
   "approve_edit.less_than_swap": "ค่าเผื่อไม่สามารถน้อยกว่าจำนวนเงินสวอปได้",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "อย่าดาวน์เกรดลงเป็น OneKey v4 การดาวน์เกรดจะทำให้ข้อมูลสูญหายแบบถาวร",
   "downgrade_warning_title": "สำคัญ: อย่าทำการดาวน์เกรด",
   "earn.24h_earnings": "รายได้ 24 ชั่วโมง",
-  "earn.24h_earnings_tooltip": "รายได้โดยประมาณในช่วง 24 ชั่วโมงที่ผ่านมาตามจำนวนที่เปิดใช้งานของคุณ",
+  "earn.24h_earnings_tooltip": "ประมาณรายได้ใน 24 ชั่วโมงตามจำนวนเงินที่คุณได้เปิดใช้งาน",
   "earn.active": "เปิดใช้งานอยู่",
   "earn.annually": "ทุกปี",
   "earn.approaching_staking_cap": "เมื่อเรากำลังเข้าใกล้ขีดจำกัดการเสนอส่วน, กรุณาตรวจสอบผลการเสนอส่วนของคุณ คุณจะต้องเสนอส่วนใหม่ถ้าถูกทำเครื่องหมายว่าเกินขีดจำกัด",
@@ -807,7 +807,7 @@
   "global.current": "ปัจจุบัน",
   "global.customize_nonce": "ปรับแต่ง nonce",
   "global.customize_nonce_desc": "เปิดสิ่งนี้เพื่อเปลี่ยน nonce (หมายเลขธุรกรรม) และควบคุมลำดับธุรกรรมของคุณบนการยืนยัน > การตั้งค่าล่วงหน้า นี่เป็นฟีเจอร์ขั้นสูง โปรดใช้ด้วยความระมัดระวัง",
-  "global.dapp_interaction": "การโต้ตอบของ Dapp",
+  "global.dapp_interaction": "การโต้ตอบของ dApp",
   "global.dark": "มืด",
   "global.date": "วันที่",
   "global.date_today": "วันนี้",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "ต้องอนุมัติโทเค็นก่อนสลับ",
   "provider.ios_popover.approval_require_title": "ต้องการการอนุมัติ",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "ใบเสนอราคารวมค่าธรรมเนียม OneKey 0.3% เพื่อให้เราสามารถให้บริการคุณภาพสูงและประสบการณ์ที่ไม่มีปัญหา",
+  "provider.ios_popover.onekey_fee_content": "ไม่มีค่าธรรมเนียมแอบแฝง OneKey คิดค่าธรรมเนียม 0.3% ซึ่งรวมอยู่ในใบเสนอราคาแล้ว โดยไม่มีค่าธรรมเนียมสำหรับการทำธุรกรรมแบบห่อหุ้ม สิ่งนี้ช่วยให้เราสามารถมอบประสบการณ์ที่มีคุณภาพสูงและราบรื่น",
   "provider.ios_popover.order_info_title": "ข้อมูลการสั่งซื้อ",
   "provider.ios_popover.title": "ข้อมูล",
   "provider.label_anti_mev": "ต่อต้าน MEV",
