+++ packages/shared/src/locale/json/uk_UA.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "Видалити контакт",
   "address_book_edit_address.title": "Редагувати адресу",
   "address_book_select.title": "Вибрати контакт",
-  "address_input.contract_popover": "Ви надсилаєте на адресу контракту, а не на звичайну адресу користувача. Ретельно перевірте адресу та стежте за ризиками.",
+  "address_input.contract_popover": "Ви надсилаєте на адресу контракту, а не на звичайну адресу користувача. Неправильні перекази на адресу контракту можуть призвести до втрати активів. Перевірте адресу двічі та зверніть увагу на ризики.",
   "address_input.first_transfer_popover": "Перший переказ на цю адресу в {network} . Ретельно перевірте адресу та стежте за ризиками.",
-  "address_input.transferred_popover": "Виявлено попередній переказ на цю адресу. Продовжуйте, але будьте обережні.",
+  "address_input.transferred_popover": "Виявлено попередні перекази на цю адресу. Не забудьте перевірити ще раз, щоб переконатися, що це потрібна адреса.",
   "address_type": "Тип адреси",
   "approve_edit.approve_amount": "Затвердити суму",
   "approve_edit.less_than_swap": "Надбавка не може бути меншою за суму обміну",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "НЕ знижуйте версію до OneKey v4. Зниження версії призведе до незворотної втрати даних.",
   "downgrade_warning_title": "Важливо: не знижуйте версію",
   "earn.24h_earnings": "Дохід за 24 години",
-  "earn.24h_earnings_tooltip": "Оціночний дохід за останні 24 години на основі вашої активної суми.",
+  "earn.24h_earnings_tooltip": "Оцінка доходу за 24 години на основі активованої суми.",
   "earn.active": "Активний",
   "earn.annually": "Щорічно",
   "earn.approaching_staking_cap": "Оскільки ми наближаємось до ліміту ставок, будь ласка, перевірте результати своїх ставок. Вам потрібно буде знову зробити ставку, якщо вона позначена як перевищення.",
@@ -807,7 +807,7 @@
   "global.current": "Поточний",
   "global.customize_nonce": "Налаштувати nonce",
   "global.customize_nonce_desc": "Увімкніть це, щоб змінити nonce (номер транзакції) і контролювати своє замовлення транзакції після підтвердження > попередні налаштування. Це розширена функція, використовуйте її обережно.",
-  "global.dapp_interaction": "Взаємодія Dapp",
+  "global.dapp_interaction": "Взаємодія dApp",
   "global.dark": "Темний",
   "global.date": "Дата",
   "global.date_today": "Сьогодні",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "Необхідно затвердити маркер перед обміном",
   "provider.ios_popover.approval_require_title": "Потрібне підтвердження",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "Цитата включає 0,3% комісії OneKey для забезпечення високоякісного обслуговування та безшовного досвіду",
+  "provider.ios_popover.onekey_fee_content": "Ніяких прихованих комісій. OneKey стягує комісію 0,3%, яка вже включена у вашу пропозицію, без додаткових зборів за обгортання транзакцій. Це підтримує нас у наданні високоякісного, безперебійного обслуговування.",
   "provider.ios_popover.order_info_title": "Інформація про замовлення",
   "provider.ios_popover.title": "Інформація",
   "provider.label_anti_mev": "Анти-МЕВ",
