+++ packages/shared/src/locale/json/vi.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "Xóa Liên hệ",
   "address_book_edit_address.title": "Chỉnh sửa Địa chỉ",
   "address_book_select.title": "Chọn Liên hệ",
-  "address_input.contract_popover": "Bạn đang gửi đến địa chỉ hợp đồng, không phải địa chỉ người dùng thông thường. Hãy kiểm tra lại địa chỉ và chú ý đến rủi ro.",
+  "address_input.contract_popover": "Bạn đang gửi đến một địa chỉ hợp đồng, không phải địa chỉ người dùng thông thường. Chuyển khoản không chính xác đến địa chỉ hợp đồng có thể dẫn đến mất tài sản. Kiểm tra kỹ địa chỉ và chú ý đến các rủi ro.",
   "address_input.first_transfer_popover": "Đầu tiên hãy chuyển đến địa chỉ này trên {network} . Kiểm tra lại địa chỉ và chú ý đến rủi ro.",
-  "address_input.transferred_popover": "Đã phát hiện chuyển khoản trước đó đến địa chỉ này. Vui lòng tiếp tục, nhưng hãy thận trọng.",
+  "address_input.transferred_popover": "Đã phát hiện các lần chuyển trước đến địa chỉ này. Đừng quên kiểm tra kỹ để đảm bảo đó là địa chỉ dự định.",
   "address_type": "Loại địa chỉ",
   "approve_edit.approve_amount": "Phê duyệt số tiền",
   "approve_edit.less_than_swap": "Khoản trợ cấp không được nhỏ hơn số tiền hoán đổi",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "ĐỪNG hạ cấp xuống OneKey v4. Việc hạ cấp sẽ dẫn đến mất dữ liệu không thể đảo ngược.",
   "downgrade_warning_title": "Quan trọng: Không nên hạ cấp",
   "earn.24h_earnings": "Thu nhập trong 24h",
-  "earn.24h_earnings_tooltip": "Thu nhập ước tính trong 24 giờ qua dựa trên số lượng hoạt động của bạn.",
+  "earn.24h_earnings_tooltip": "Thu nhập ước tính trong 24 giờ dựa trên số tiền đã kích hoạt của bạn.",
   "earn.active": "Đang hoạt động",
   "earn.annually": "Hàng năm",
   "earn.approaching_staking_cap": "Khi chúng ta đang tiếp cận giới hạn staking, vui lòng kiểm tra kết quả staking của bạn. Bạn sẽ cần restake nếu được đánh dấu là overflow.",
@@ -807,7 +807,7 @@
   "global.current": "Hiện tại",
   "global.customize_nonce": "Tùy chỉnh nonce",
   "global.customize_nonce_desc": "Bật tính năng này để thay đổi nonce (số giao dịch) và kiểm soát lệnh giao dịch của bạn khi xác nhận > cài đặt nâng cao. Đây là tính năng nâng cao, hãy sử dụng thận trọng.",
-  "global.dapp_interaction": "Tương tác Dapp",
+  "global.dapp_interaction": "Tương tác dApp",
   "global.dark": "Tối tăm",
   "global.date": "Ngày",
   "global.date_today": "Hôm nay",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "Cần phải chấp thuận mã thông báo trước khi hoán đổi",
   "provider.ios_popover.approval_require_title": "Yêu cầu phê duyệt",
   "provider.ios_popover.onekey_fee": "OneKey fee",
-  "provider.ios_popover.onekey_fee_content": "Báo giá bao gồm phí OneKey 0,3% để đảm bảo chúng tôi cung cấp dịch vụ chất lượng cao và trải nghiệm mượt mà.",
+  "provider.ios_popover.onekey_fee_content": "Không có phí ẩn. OneKey tính phí 0,3%, đã bao gồm trong báo giá của bạn, không có phí cho các giao dịch bọc. Điều này hỗ trợ chúng tôi cung cấp trải nghiệm liền mạch, chất lượng cao.",
   "provider.ios_popover.order_info_title": "Thông tin Đơn hàng",
   "provider.ios_popover.title": "Thông tin",
   "provider.label_anti_mev": "Chống MEV",
