+++ packages/shared/src/locale/json/zh_CN.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "删除联系人",
   "address_book_edit_address.title": "编辑地址",
   "address_book_select.title": "选择联系人",
-  "address_input.contract_popover": "您正在发送到合约地址，而非常规用户地址。请仔细检查地址并注意风险。",
+  "address_input.contract_popover": "您正在发送到合约地址，而不是普通用户地址。错误地转账到合约地址可能导致资产损失。请仔细检查地址并注意风险。",
   "address_input.first_transfer_popover": "首次在 {network} 上向此地址转账。仔细核对地址并注意风险。",
-  "address_input.transferred_popover": "检测到之前有向此地址的转账",
+  "address_input.transferred_popover": "已检测到之前向此地址的转账。也请再次检查以确保这是预期的地址。",
   "address_type": "地址类型",
   "approve_edit.approve_amount": "授权金额",
   "approve_edit.less_than_swap": "授权额度不能低于兑换金额",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "请勿降级至 OneKey v4。降级将导致无法恢复的数据丢失。",
   "downgrade_warning_title": "重要：不要降级",
   "earn.24h_earnings": "24 小时收益",
-  "earn.24h_earnings_tooltip": "根据您已激活的数量，过去 24 小时的估计收益。",
+  "earn.24h_earnings_tooltip": "根据您已激活的数量所预估的 24 小时收益。",
   "earn.active": "已激活",
   "earn.annually": "每年",
   "earn.approaching_staking_cap": "由于我们正在接近质押上限，请检查您的质押结果。如果标记为溢出，您将需要重新质押。",
@@ -378,7 +378,7 @@
   "earn.native_staking": "原生质押",
   "earn.no_orders": "无订单",
   "earn.no_orders_desc": "您还没有质押任何资产。",
-  "earn.not_enough_token": "不足 {token}？",
+  "earn.not_enough_token": "{token} 不足？",
   "earn.number_day": "{number} 天",
   "earn.number_days": "{number} 天",
   "earn.number_days_number_block": "{number_days} 天 ({number} 个区块)",
@@ -767,9 +767,9 @@
   "global.circulating_supply": "流通供应量",
   "global.clear": "清除",
   "global.close": "关闭",
-  "global.collect_to_device": "收集到设备",
+  "global.collect_to_device": "收藏到设备",
   "global.collect_to_device_canceled": "已在设备上取消",
-  "global.collect_to_device_failed": "收集失败，请再试一次",
+  "global.collect_to_device_failed": "收藏失败，请再试一次",
   "global.community": "社区",
   "global.confirm": "确认",
   "global.confirm_on_device": "在设备上确认",
@@ -807,7 +807,7 @@
   "global.current": "当前",
   "global.customize_nonce": "自定义随机数",
   "global.customize_nonce_desc": "启用此功能可更改 nonce（随机数，又为交易编号）并在交易确认 > 高级设置时控制交易顺序。这是一项高级功能，请谨慎使用。",
-  "global.dapp_interaction": "Dapp 交互",
+  "global.dapp_interaction": "dApp 交互",
   "global.dark": "深色",
   "global.date": "日期",
   "global.date_today": "今天",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "兑换前需授权代币",
   "provider.ios_popover.approval_require_title": "授权",
   "provider.ios_popover.onekey_fee": "OneKey 费用",
-  "provider.ios_popover.onekey_fee_content": "报价包含 0.3% 的 OneKey 费用，以确保我们提供高质量的服务和无缝的体验",
+  "provider.ios_popover.onekey_fee_content": "没有隐藏费用。OneKey 收取0.3%的费用，已包含在您的报价中，Wrapped token 时不收取費用。这将支持我们提供高质量、无缝的体验。",
   "provider.ios_popover.order_info_title": "订单信息",
   "provider.ios_popover.title": "提示",
   "provider.label_anti_mev": "防MEV",
