+++ packages/shared/src/locale/json/zh_TW.json
@@ -52,9 +52,9 @@
   "address_book_edit_address.delete_contact_title": "刪除聯絡人",
   "address_book_edit_address.title": "編輯地址",
   "address_book_select.title": "選擇聯絡人",
-  "address_input.contract_popover": "您正在發送到合約地址，而不是常規用戶地址。仔細檢查地址並注意風險。",
+  "address_input.contract_popover": "您正在發送到合約地址，而不是普通用戶地址。錯誤地轉賬到合約地址可能會導致資產損失。請仔細檢查地址並注意風險。",
   "address_input.first_transfer_popover": "首次在 {network} 上向此地址轉賬。仔細核对地址並注意風險。",
-  "address_input.transferred_popover": "檢測到之前有向此地址的轉賬",
+  "address_input.transferred_popover": "已檢測到之前向此地址的轉帳。也請再次檢查以確保這是預期的地址。",
   "address_type": "地址類型",
   "approve_edit.approve_amount": "授權金額",
   "approve_edit.less_than_swap": "授權額度不能低於兌換金額",
@@ -313,7 +313,7 @@
   "downgrade_warning_description": "請勿降級至 OneKey v4。降級將導致無法逆轉的資料損失。",
   "downgrade_warning_title": "重要：請勿降級",
   "earn.24h_earnings": "24 小時收益",
-  "earn.24h_earnings_tooltip": "根據您已啟用的數量，過去 24 小時的估計收益。",
+  "earn.24h_earnings_tooltip": "根據您已啟用的數量所預估的 24 小時收益。",
   "earn.active": "已啟用",
   "earn.annually": "每年",
   "earn.approaching_staking_cap": "由於我們即將接近質押上限，請檢查您的質押結果。如果標記為溢出，您將需要重新質押。",
@@ -378,7 +378,7 @@
   "earn.native_staking": "原生質押",
   "earn.no_orders": "無訂單",
   "earn.no_orders_desc": "您尚未質押任何資產。",
-  "earn.not_enough_token": "不足 {token}？",
+  "earn.not_enough_token": "{token} 不足？",
   "earn.number_day": "{number} 天",
   "earn.number_days": "{number} 天",
   "earn.number_days_number_block": "{number_days} 天 ({number} 個區塊)",
@@ -767,9 +767,9 @@
   "global.circulating_supply": "流通供應量",
   "global.clear": "清除",
   "global.close": "關閉",
-  "global.collect_to_device": "收集到裝置",
+  "global.collect_to_device": "收藏到裝置",
   "global.collect_to_device_canceled": "已在裝置上取消",
-  "global.collect_to_device_failed": "收集失敗，請再試一次",
+  "global.collect_to_device_failed": "收藏失敗，請再試一次",
   "global.community": "社群",
   "global.confirm": "確認",
   "global.confirm_on_device": "在裝置上確認",
@@ -807,7 +807,7 @@
   "global.current": "當前",
   "global.customize_nonce": "自訂隨機數",
   "global.customize_nonce_desc": "開啟此選項可變更 nonce（隨機數，又為交易編號）並在交易確認 > 進階設定中控制您的交易訂單。這是一項高級功能，請謹慎使用。",
-  "global.dapp_interaction": "Dapp 交互",
+  "global.dapp_interaction": "dApp 交互",
   "global.dark": "深色",
   "global.date": "日期",
   "global.date_today": "今天",
@@ -1419,7 +1419,7 @@
   "provider.ios_popover.approval_require_msg": "兌換前需要授權代幣",
   "provider.ios_popover.approval_require_title": "授權",
   "provider.ios_popover.onekey_fee": "OneKey 費用",
-  "provider.ios_popover.onekey_fee_content": "報價包含 0.3% 的 OneKey 費用，以確保我們提供高品質的服務和無縫的體驗",
+  "provider.ios_popover.onekey_fee_content": "沒有隱藏費用。OneKey 收取 0.3% 的費用，已包含在您的報價中，Wrapped token 時不收取費用。這支持我們提供高品質、無縫的體驗。",
   "provider.ios_popover.order_info_title": "訂單資訊",
   "provider.ios_popover.title": "提示",
   "provider.label_anti_mev": "防MEV",
